feat(sdk): run TradingAgents artifacts on ATL
