+++ packaging/agentictrading/src/agentictrading/integrations/__init__.py
@@ -6,9 +6,12 @@ from .tradingagents import (
     TradingAgentsDecisionArtifact,
     TradingAgentsDecisionGenerator,
     TradingAgentsDecisionRecord,
+    TradingAgentsATLRunner,
+    TradingAgentsATLRunOutcome,
     TradingAgentsDependencyError,
     TradingAgentsGenerationError,
     TradingAgentsReplayDiagnostics,
+    TradingAgentsReplayIncompleteError,
     TradingAgentsReplayPlanner,
     TradingAgentsReplayValidationError,
     TradingAgentsVersionError,
@@ -25,9 +28,12 @@ __all__ = [
     "TradingAgentsDecisionArtifact",
     "TradingAgentsDecisionGenerator",
     "TradingAgentsDecisionRecord",
+    "TradingAgentsATLRunner",
+    "TradingAgentsATLRunOutcome",
     "TradingAgentsDependencyError",
     "TradingAgentsGenerationError",
     "TradingAgentsReplayDiagnostics",
+    "TradingAgentsReplayIncompleteError",
     "TradingAgentsReplayPlanner",
     "TradingAgentsReplayValidationError",
     "TradingAgentsVersionError",
